+++ final/Web/service/src/main/java/com/expexchangeservice/service/interfaces/IUserProfileService.java
@@ -12,13 +12,13 @@ import java.util.List;
 public interface IUserProfileService {
 
     void addUserProfile(ProfileDto profileDto);
-    boolean changeUserProfile(int profileId, ProfileDto profileDto);
-    boolean deleteUserProfile(int profileId);
+    boolean changeUserProfile(Long profileId, ProfileDto profileDto);
+    boolean deleteUserProfile(Long profileId);
     ProfileDto getProfileDtoByUsername(String username);
     UserProfile findProfileByUsername(String username);
     boolean signUpForTheLesson(String username, Lesson lesson);
     boolean signUpForTheCourse(String username, Course course);
-    ProfileDto getUserProfileById(Integer profileId);
+    ProfileDto getUserProfileById(Long profileId);
     List<LessonDto> getLessonListForUser(String username);
     List<CourseDto> getCourseListForUser(String username);
     boolean changeUserRole(String username, boolean isAdmin);
